safari-600.7-branch/LayoutTests/fast/css/overflow-property-expected.txt

This tests the parsing and retrieval of the overflow, overflow-x and overflow-y properties.

PASS test0.style.overflow is "auto"
PASS test0.style.overflowX is "auto"
PASS test0.style.overflowY is "auto"
PASS test0.style.getPropertyShorthand('overflow-x') is "overflow"
PASS test0.style.getPropertyShorthand('overflow-y') is "overflow"
PASS test1.style.overflow is ""
PASS test1.style.overflowX is ""
PASS test1.style.overflowY is ""
PASS test2.style.overflow is "scroll"
PASS test2.style.overflowX is "scroll"
PASS test2.style.overflowY is "scroll"
PASS test3.style.overflow is ""
PASS test3.style.overflowX is "overlay"
PASS test3.style.overflowY is "hidden"
PASS test4.style.overflow is ""
PASS test4.style.overflowX is "auto"
PASS test4.style.overflowY is ""
PASS test5.style.overflow is ""
PASS test5.style.overflowX is ""
PASS test5.style.overflowY is "auto"
PASS test6.style.overflow is ""
PASS test6.style.overflowX is ""
PASS test6.style.overflowY is ""
PASS test7.style.overflow is ""
PASS test7.style.overflowX is ""
PASS test7.style.overflowY is ""
PASS test8.style.overflowX is ""
PASS test8.style.overflowY is "-webkit-paged-x"
PASS test9.style.overflowX is "auto"
PASS test9.style.overflowY is "-webkit-paged-y"
PASS test10.style.overflowX is "visible"
PASS test10.style.overflowY is "-webkit-paged-y"
PASS successfullyParsed is true

TEST COMPLETE
